requirements.txt
xgboost
scikit-learn
mlflow
pandas
numpy
psycopg2-binary